media-auto-ia/automotive-media-engine/requirements.txt
# Zero-Cost Core
google-generativeai>=0.8.0
edge-tts>=6.1.9
manim>=0.18.0
duckduckgo-search>=5.0.0

# Legacy / Optional (Paid)
anthropic>=0.40.0
elevenlabs>=1.0.0

fastmcp>=0.2.0
pydantic>=2.0.0
python-dotenv>=1.0.0
Pillow>=10.0.0
matplotlib>=3.8.0
requests>=2.31.0
PyYAML>=6.0.1
pytest>=7.4.0
black>=23.0.0
ruff>=0.1.0

google-api-python-client
google-auth-oauthlib
google-auth-httplib2
elevenlabs==0.2.27
